+++ usr/local/bin/autonfs.sh
@@ -22,6 +22,7 @@ fi
 # The init script handles the kill process and if there`s already a process running
 echo $$ > /var/run/autonfs.pid
 
+declare -a MNTPATH
 while true; do
   # First check if Fileserver is serving NFS
   /usr/bin/rpcinfo -t "$FILESERVER" nfs $NFSVERS &>/dev/null
@@ -34,21 +35,33 @@ while true; do
       # Second check was succssful, checking if all NFS Shares are mounted
       logger -t autonfs "Fileserver[${FILESERVER}] is up"
       for MOUNT in ${MOUNTS[@]}; do
-        mount | grep -E "^${FILESERVER}:${MOUNT} on .*$" &>/dev/null
+        # Split the share if it has a different remote and local mountpoint
+        # Both will be the same if the mountpoints are so too
+        MNTPATH=(`echo ${MOUNT//$MOUNTSDELIMITER/ }`)
+        RMNTPATH=${MNTPATH[0]}
+        LMNTPATH=${MNTPATH[${#MNTPATH[@]}-1]}
+        # Let's check if its already mounted
+        mount | grep -E "^${FILESERVER}:${RMNTPATH} on .*$" &>/dev/null
         if [ $? -ne 0 ]; then
           # The actual NFS Share is not mounted, so it will be done now
           #logger -t autonfs "NFS-Share[${MOUNT}] not mounted - attempting to mount it from: ${FILESERVER}:${MOUNT}"
-          mount -t nfs ${MOUNTOPT} ${FILESERVER}:${MOUNT} ${MOUNT}
+          mount -t nfs ${MOUNTOPT} ${FILESERVER}:${RMNTPATH} ${LMNTPATH}
         fi
       done
     else
       # The Fileserver was not reachable two times, unmounting shares which are still mounted
       logger -t autonfs "Fileserver(${FILESERVER}) is down."
       for MOUNT in ${MOUNTS[@]}; do
-        mount | grep -E "^${FILESERVER}:${MOUNT} on .*$" &>/dev/null
+        # Split the share if it has a different remote and local mountpoint
+        # Both will be the same if the mountpoints are so too
+        MNTPATH=(`echo ${MOUNT//$MOUNTSDELIMITER/ }`)
+        RMNTPATH=${MNTPATH[0]}
+        LMNTPATH=${MNTPATH[${#MNTPATH[@]}-1]}
+        # Let's check if its already mounted
+        mount | grep -E "^${FILESERVER}:${RMNTPATH} on .*$" &>/dev/null
         if [ $? -eq 0 ]; then
-          logger -t autonfs "Cannot reach ${FILESERVER}, therefore unmounting NFS-Share[${MOUNT}]"
-          umount -f ${MOUNT}
+          logger -t autonfs "Cannot reach ${FILESERVER}, therefore unmounting NFS-Share[${LMNTPATH}]"
+          umount -f ${LMNTPATH}
         fi
       done
     fi
@@ -56,11 +69,17 @@ while true; do
     # First check succeeded, so just checking if all NFS-Shares are mounted
     #logger -t autonfs "Fileserver[${FILESERVER}] is up"
     for MOUNT in ${MOUNTS[@]}; do
-      mount | grep -E "^${FILESERVER}:${MOUNT} on .*$" &>/dev/null
+      # Split the share if it has a different remote and local mountpoint
+      # Both will be the same if the mountpoints are so too
+      MNTPATH=(`echo ${MOUNT//$MOUNTSDELIMITER/ }`)
+      RMNTPATH=${MNTPATH[0]}
+      LMNTPATH=${MNTPATH[${#MNTPATH[@]}-1]}
+      # Let's check if its already mounted
+      mount | grep -E "^${FILESERVER}:${RMNTPATH} on .*$" &>/dev/null
       if [ $? -ne 0 ]; then
         # The actual NFS Share is not mounted, so it will be done now
         #logger -t autonfs "NFS-Share[${MOUNT}] not mounted - attempting to mount it from: ${FILESERVER}:${MOUNT}"
-        mount -t nfs ${MOUNTOPT} ${FILESERVER}:${MOUNT} ${MOUNT}
+        mount -t nfs ${MOUNTOPT} ${FILESERVER}:${RMNTPATH} ${LMNTPATH}
       fi
     done
   fi